[ita-stf] StatePayload: change `state_update` field to type `SgxStateDiffType`
