+++ app-libs/stf/Cargo.toml
@@ -21,7 +21,6 @@ std = [
     "log",
     "base58",
     "sc-keystore",
-    "serde",
     "system/std",
     "sp-core/std",
     "hex",
@@ -42,7 +41,6 @@ log = { version = "0.4", optional = true }
 base58 = { version = "0.1", optional = true }
 derive_more = { version = "0.99.5", optional = true }
 hex = { version = "0.4.2", optional = true }
-serde = { version = "1.0", features = ["derive"], optional = true }
 codec = { version = "2.0.0", default-features = false, features = ["derive"], package = "parity-scale-codec" }
 sgx_tstd = { rev = "v1.1.3", features = ["untrusted_fs","net","backtrace"], git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
