+++ app-libs/stf/src/stf_sgx_primitives.rs
@@ -19,11 +19,62 @@ use crate::{AccountId, AccountInfo, Index, ShardIdentifier};
 use codec::{Decode, Encode};
 use derive_more::Display;
 use itp_storage::{storage_map_key, StorageHasher};
+use itp_types::H256;
 use log_sgx::*;
 use sgx_tstd as std;
 use std::{prelude::v1::*, vec};
 
 pub type StfResult<T> = Result<T, StfError>;
+
+pub mod types {
+	pub use sgx_runtime::{Balance, Index};
+	pub type AccountData = balances::AccountData<Balance>;
+	pub type AccountInfo = system::AccountInfo<Index, AccountData>;
+
+	pub type StateType = sgx_externalities::SgxExternalitiesType;
+	pub type State = sgx_externalities::SgxExternalities;
+	pub type StateTypeDiff = sgx_externalities::SgxExternalitiesDiffType;
+	pub use super::StatePayload;
+	pub struct Stf;
+}
+
+use types::StateTypeDiff;
+
+/// payload of block that needs to be encrypted
+#[derive(PartialEq, Eq, Clone, Encode, Decode, Debug)]
+pub struct StatePayload {
+	/// state hash before the `state_update` was applied.
+	state_hash_apriori: H256,
+	/// state hash after the `state_update` was applied.
+	state_hash_aposteriori: H256,
+	/// state diff applied to state with hash `state_hash_apriori`
+	/// leading to state with hash `state_hash_aposteriori`
+	state_update: StateTypeDiff,
+}
+
+impl StatePayload {
+	/// get state hash before the `state_update` was applied.
+	pub fn state_hash_apriori(&self) -> H256 {
+		self.state_hash_apriori
+	}
+	/// get state hash after the `state_update` was applied.
+	pub fn state_hash_aposteriori(&self) -> H256 {
+		self.state_hash_aposteriori
+	}
+	/// reference to the `state_update`
+	pub fn state_update(&self) -> &StateTypeDiff {
+		&self.state_update
+	}
+
+	/// create new `StatePayload` instance.
+	pub fn new(apriori: H256, aposteriori: H256, update: StateTypeDiff) -> StatePayload {
+		StatePayload {
+			state_hash_apriori: apriori,
+			state_hash_aposteriori: aposteriori,
+			state_update: update,
+		}
+	}
+}
 
 #[derive(Debug, Display, PartialEq, Eq)]
 pub enum StfError {
